+++ packages/react-select/src/index.ts
@@ -41,7 +41,7 @@ export type { OptionProps } from './components/Option';
 export type { PlaceholderProps } from './components/Placeholder';
 export type { SingleValueProps } from './components/SingleValue';
 export type { ThemeConfig } from './theme';
-export type { StylesConfig } from './styles';
+export type { ClassNamesConfig, StylesConfig } from './styles';
 export * from './types';
 export type {
   OptionContext,